Course Assignment 1
https://github.com/Lukusmanucus/prog-found-ca.git

Sorry for the late upload. Spent about 45 min admiring my work on my accomplishment of figuring out question 10 (and playing with it) without web-help (and arrived late for work).
Also found out how cool it is adding a loop to the assignment even though it was not called for..... hope that does not give a negative impact on my assessment. I'll work on lvl 2 later this evening and deliver that if I have time.